src/embeddings/requirements.txt
# langchain-openai
langchain
# openai==1.10.0
langchain-openai
python-dotenv
langchain_community